Fixing minor issues
