--- html/src/layouts/default/Default.vue
@@ -1,9 +1,0 @@
-<template>
-  <v-app full-height>
-    <default-view />
-  </v-app>
-</template>
-
-<script lang="ts" setup>
-  import DefaultView from './View.vue'
-</script>
