--- html/src/layouts/default/View.vue
@@ -1,3 +1,0 @@
-<template>
-  <router-view />
-</template>
