+++ html/src/main.ts
@@ -11,7 +11,7 @@ import App from './App.vue'
 import { createApp } from 'vue'
 
 // Plugins
-import { registerPlugins } from '@/plugins'
+import { registerPlugins } from './plugins'
 
 const app = createApp(App)
