+++ html/src/plugins/index.ts
@@ -5,18 +5,14 @@
  */
 
 // Plugins
-import { loadFonts } from './webfontloader'
-import vuetify from './vuetify'
-import pinia from '../store'
-import router from '../router'
+import { loadFonts } from "./webfontloader";
+import vuetify from "./vuetify";
+import pinia from "../store";
 
 // Types
-import type { App } from 'vue'
+import type { App } from "vue";
 
-export function registerPlugins (app: App) {
-  loadFonts()
-  app
-    .use(vuetify)
-    .use(router)
-    .use(pinia)
+export function registerPlugins(app: App) {
+  loadFonts();
+  app.use(vuetify).use(pinia);
 }
