+++ html/src/plugins/vuetify.ts
@@ -14,7 +14,6 @@ import { createVuetify } from 'vuetify'
 // https://vuetifyjs.com/en/introduction/why-vuetify/#feature-guides
 export default createVuetify({
   theme: {
-    defaultTheme: 'dark',
     themes: {
       dark: {
         colors: {
