+++ html/src/store/global.ts
@@ -3,11 +3,6 @@ import { defineStore } from 'pinia'
 
 export const useGlobalStore = defineStore('global', {
   state: () => ({
-    playerID: "123"
-  }),
-  getters: {
-    playerID(state) {
-      return state.playerID
-    }
-  }
+    playerID: ""
+  })
 })
